+++ CHANGELOG.rst
@@ -2,9 +2,10 @@
 =============
 
 - Fix incompatibility with ``multiprocessing`` on Windows by entirely refactoring the internal structure of the ``logger`` so it can be inherited by child processes along with added handlers (`#108 <https://github.com/Delgan/loguru/issues/108>`_).
+- Fix an error using a ``filter`` function "by name" while receiving a log with ``record["name"]`` equals to ``None``.
 - Add support for ``copy.deepcopy()`` of the ``logger`` allowing multiple independent loggers with separate set of handlers (`#72 <https://github.com/Delgan/loguru/issues/72>`_).
 - Prevent hypothetical ``ImportError`` if a Python installation is missing the built-in ``distutils`` module (`#118 <https://github.com/Delgan/loguru/issues/118>`_).
-- Fix an error using a ``filter`` function "by name" while receiving a log with ``record["name"]`` equals to ``None``.
+- Add the level ``name`` as the first argument of namedtuple returned by the ``.level()`` method.
 
 
 `0.3.2`_ (2019-07-21)